pyproject.toml
[project]
name = "speech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "mutagen>=1.47.0",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "torch-summary>=1.4.5",
]
